graduate C3 React (Workers) template
Resolves DEVX-1697
