+++ .changeset/mighty-moments-march.md
@@ -1,0 +1,5 @@
+---
+"create-cloudflare": patch
+---
+
+graduate C3 React (Workers) template
