+++ packages/create-cloudflare/e2e-tests/frameworks/framework-test-config.ts
@@ -31,7 +31,7 @@ export default function getFrameworkTestConfig(pm: string) {
 				"strict",
 			],
 		},
-		["docusaurus:pages"]: {
+		"docusaurus:pages": {
 			argv: ["--platform", "pages"],
 			unsupportedPms: ["bun"],
 			testCommitMessage: true,
@@ -57,7 +57,7 @@ export default function getFrameworkTestConfig(pm: string) {
 				},
 			],
 		},
-		["docusaurus:workers"]: {
+		"docusaurus:workers": {
 			argv: ["--platform", "workers"],
 			unsupportedPms: ["bun"],
 			testCommitMessage: true,
@@ -265,7 +265,8 @@ export default function getFrameworkTestConfig(pm: string) {
 				expectedText: "C3_TEST",
 			},
 		},
-		react: {
+		"react:pages": {
+			argv: ["--platform", "pages"],
 			promptHandlers: [
 				{
 					matcher: /Select a variant:/,
@@ -282,6 +283,34 @@ export default function getFrameworkTestConfig(pm: string) {
 			verifyPreview: {
 				route: "/",
 				expectedText: "Vite + React",
+			},
+		},
+		"react:workers": {
+			argv: ["--platform", "workers"],
+			promptHandlers: [
+				{
+					matcher: /Select a variant:/,
+					input: [keys.enter],
+				},
+			],
+			unsupportedOSs: ["win32"],
+			testCommitMessage: true,
+			verifyDeploy: {
+				route: "/",
+				// Note that this is the text in the static HTML that is returned
+				// This React SPA will change this at runtime but we are only making a fetch request
+				// not actually running the client side JS.
+				expectedText: "Vite + React + TS",
+			},
+			verifyPreview: {
+				route: "/",
+				// We need to run the preview on the specific IP address on which we make the request.
+				// By default `vite preview` runs on `localhost` that doesn't always include 127.0.0.1.
+				previewArgs: ["--host=127.0.0.1"],
+				// Note that this is the text in the static HTML that is returned
+				// This React SPA will change this at runtime but we are only making a fetch request
+				// not actually running the client side JS.
+				expectedText: "Vite + React + TS",
 			},
 		},
 		solid: {
