+++ packages/create-cloudflare/templates/react/workers/js/api/index.js
