+++ packages/create-cloudflare/templates/react/workers/js/src/App.jsx
