+++ packages/create-cloudflare/templates/react/workers/js/wrangler.jsonc
