+++ packages/create-cloudflare/templates/react/workers/ts/src/App.tsx
