+++ packages/create-cloudflare/templates/react/workers/ts/tsconfig.worker.json
