+++ packages/create-cloudflare/templates/react/workers/ts/worker-configuration.d.ts
