+++ packages/create-cloudflare/templates/react/workers/ts/wrangler.jsonc
